Move filtering by contact stuff to transactions module
