+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/transactions/TransactionFilterService.kt
@@ -1,7 +1,6 @@
 package io.horizontalsystems.bankwallet.modules.transactions
 
 import io.horizontalsystems.bankwallet.entities.Wallet
-import io.horizontalsystems.bankwallet.modules.contacts.SelectContactViewModel
 import io.horizontalsystems.bankwallet.modules.contacts.model.Contact
 import io.horizontalsystems.marketkit.models.Blockchain
 import kotlinx.coroutines.flow.MutableStateFlow
